requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI and NLP
google-generativeai==0.3.2
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
pyodbc==5.0.1

# Scheduling
apscheduler==3.10.4

# Social Media APIs
linkedin-api==2.0.0
tweepy==4.14.0

# HTTP and API
requests==2.31.0

# Utilities
schedule==1.2.0

# Additional dependencies
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1 